Just a little bit of BT17 updating
Also unrelated to that but Theodoro has a mission that changes their culture to the Germanic group in vanilla, BT16 CI never adapted this new culture but now I've decided to make a new East Germanic group for them. As Theodoro starts out as Pontic instead of Gothic in CI this is a bit weird but honestly the mission in vanilla is quite dumb anyway.